+++ target/disable.ps1
@@ -61,30 +61,45 @@ try {
     }
 
     # Process
-    switch ($action) {
-        'DisableAccount' {
-            if (-not($actionContext.DryRun -eq $true)) {
-                Write-Information "Disabling {connectorName} account with accountReference: [$($actionContext.References.Account)]"
-                # < Write Disable logic here >
+    switch ($actionContext.Origin) {
+        'enforcement' {
+            switch ($action) {
+                'DisableAccount' {
+                    if (-not($actionContext.DryRun -eq $true)) {
+                        Write-Information "Disabling {connectorName} account with accountReference: [$($actionContext.References.Account)]. Action initiated by: [$($actionContext.Origin)]"
+                        # < Write disable logic here >Or
+                    } else {
+                        Write-Information "[DryRun] Disable {connectorName} account with accountReference: [$($actionContext.References.Account)], will be executed during enforcement"
+                    }
 
-            } else {
-                Write-Information "[DryRun] Disable {connectorName} account with accountReference: [$($actionContext.References.Account)], will be executed during enforcement"
+                    # Make sure to filter out arrays from $outputContext.Data (If this is not mapped to type Array in the fieldmapping). This is not supported by HelloID.
+                    $outputContext.Success = $true
+                    $outputContext.AuditLogs.Add([PSCustomObject]@{
+                            Message = "Disable account was sucessful. Action initiated by: [$($actionContext.Origin)]"
+                            IsError = $false
+                        })
+                    break
+                }
+
+                'NotFound' {
+                    Write-Information "{connectorName} account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted. Action initiated by: [$($actionContext.Origin)]"
+                    $outputContext.Success = $false
+                    $outputContext.AuditLogs.Add([PSCustomObject]@{
+                            Message = "{connectorName} account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted. Action initiated by: [$($actionContext.Origin)]"
+                            IsError = $true
+                        })
+                    break
+                }
             }
-
-            # Make sure to filter out arrays from $outputContext.Data (If this is not mapped to type Array in the fieldmapping). This is not supported by HelloID.
-            $outputContext.Success = $true
-            $outputContext.AuditLogs.Add([PSCustomObject]@{
-                    Message = 'Disable account was successful'
-                    IsError = $false
-                })
-            break
         }
 
-        'NotFound' {
-            Write-Information "{connectorName} account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted"
+        'reconciliation' {
+            Write-Information "Disabling {connectorName} account with accountReference: [$($actionContext.References.Account)]. Action initiated by: [$($actionContext.Origin)]"
+            # < Write reconciliation disable logic here >
+
             $outputContext.Success = $true
             $outputContext.AuditLogs.Add([PSCustomObject]@{
-                    Message = "{connectorName} account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted"
+                    Message = "Disable account was sucessful. Action initiated by: [$($actionContext.Origin)]"
                     IsError = $false
                 })
             break
@@ -96,11 +111,11 @@ try {
     if ($($ex.Exception.GetType().FullName -eq 'Microsoft.PowerShell.Commands.HttpResponseException') -or
         $($ex.Exception.GetType().FullName -eq 'System.Net.WebException')) {
         $errorObj = Resolve-{connectorName}Error -ErrorObject $ex
-        $auditMessage = "Could not disable {connectorName} account. Error: $($errorObj.FriendlyMessage)"
-        Write-Warning "Error at Line '$($errorObj.ScriptLineNumber)': $($errorObj.Line). Error: $($errorObj.ErrorDetails)"
+        $auditMessage = "Could not disable {connectorName} account. Error: $($errorObj.FriendlyMessage). Action initiated by: [$($actionContext.Origin)]"
+        Write-Warning "Error at Line '$($errorObj.ScriptLineNumber)': $($errorObj.Line). Error: $($errorObj.ErrorDetails). Action initiated by: [$($actionContext.Origin)]"
     } else {
-        $auditMessage = "Could not disable {connectorName} account. Error: $($_.Exception.Message)"
-        Write-Warning "Error at Line '$($ex.InvocationInfo.ScriptLineNumber)': $($ex.InvocationInfo.Line). Error: $($ex.Exception.Message)"
+        $auditMessage = "Could not disable {connectorName} account. Error: $($_.Exception.Message). Action initiated by: [$($actionContext.Origin)]"
+        Write-Warning "Error at Line '$($ex.InvocationInfo.ScriptLineNumber)': $($ex.InvocationInfo.Line). Error: $($ex.Exception.Message). Action initiated by: [$($actionContext.Origin)]"
     }
     $outputContext.AuditLogs.Add([PSCustomObject]@{
         Message = $auditMessage