+++ target/import.ps1
@@ -1,0 +1,82 @@
+#################################################
+# HelloID-Conn-Prov-Target-{connectorName]-Import
+# PowerShell V2
+#################################################
+
+# Enable TLS1.2
+[System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor [System.Net.SecurityProtocolType]::Tls12
+
+#region functions
+function Resolve-{connectorName}Error {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)]
+        [object]
+        $ErrorObject
+    )
+    process {
+        $httpErrorObj = [PSCustomObject]@{
+            ScriptLineNumber = $ErrorObject.InvocationInfo.ScriptLineNumber
+            Line             = $ErrorObject.InvocationInfo.Line
+            ErrorDetails     = $ErrorObject.Exception.Message
+            FriendlyMessage  = $ErrorObject.Exception.Message
+        }
+        if (-not [string]::IsNullOrEmpty($ErrorObject.ErrorDetails.Message)) {
+            $httpErrorObj.ErrorDetails = $ErrorObject.ErrorDetails.Message
+        } elseif ($ErrorObject.Exception.GetType().FullName -eq 'System.Net.WebException') {
+            if ($null -ne $ErrorObject.Exception.Response) {
+                $streamReaderResponse = [System.IO.StreamReader]::new($ErrorObject.Exception.Response.GetResponseStream()).ReadToEnd()
+                if (-not [string]::IsNullOrEmpty($streamReaderResponse)) {
+                    $httpErrorObj.ErrorDetails = $streamReaderResponse
+                }
+            }
+        }
+        try {
+            $errorDetailsObject = ($httpErrorObj.ErrorDetails | ConvertFrom-Json)
+            # Make sure to inspect the error result object and add only the error message as a FriendlyMessage.
+            # $httpErrorObj.FriendlyMessage = $errorDetailsObject.message
+            $httpErrorObj.FriendlyMessage = $httpErrorObj.ErrorDetails # Temporarily assignment
+        } catch {
+            $httpErrorObj.FriendlyMessage = "Error: [$($httpErrorObj.ErrorDetails)] [$($_.Exception.Message)]"
+        }
+        Write-Output $httpErrorObj
+    }
+}
+#endregion
+
+try {
+    Write-Information 'Starting {connectorName} account entitlement import'
+    $splatImportAccountParams = @{
+        Uri     = '<URL to API>'
+        Method  = 'GET'
+    }
+    $importedAccounts = Invoke-RestMethod @splatGetAllAccountParams
+
+    foreach ($importedAccount in $importedAccounts) {
+        $data = @{}
+        foreach ($field in $actionContext.ImportFields) {
+            $data[$field] = $importedAccount[$field]
+        }
+
+        # Return the result
+        Write-Output @{
+            AccountReference = $importedAccount.ExternalId
+            DisplayName      = $importedAccount.DisplayName
+            UserName         = $importedAccount.UserName
+            Enabled          = $true
+            Data             = $data
+        }
+    }
+    Write-Information '{connectorName} account entitlement import completed'
+} catch {
+    $ex = $PSItem
+    if ($($ex.Exception.GetType().FullName -eq 'Microsoft.PowerShell.Commands.HttpResponseException') -or
+        $($ex.Exception.GetType().FullName -eq 'System.Net.WebException')) {
+        $errorObj = Resolve-GoogleWSError -ErrorObject $ex
+        Write-Warning "Error at Line '$($errorObj.ScriptLineNumber)': $($errorObj.Line). Error: $($errorObj.ErrorDetails)"
+        Write-Error "Could not import {connectorName} account entitlements. Error: $($errorObj.FriendlyMessage)"
+    } else {
+        Write-Warning "Error at Line '$($ex.InvocationInfo.ScriptLineNumber)': $($ex.InvocationInfo.Line). Error: $($ex.Exception.Message)"
+        Write-Error "Could not import {connectorName} account entitlements. Error: $($ex.Exception.Message)"
+    }
+}